+++ src/ClientProcess.java
@@ -13,18 +13,13 @@ import java.util.Scanner;
 public class ClientProcess {
 	private int leader = 0;
 	private String ipAddress = "";
-	private Integer port = 5000;
+	private Integer port = 5001;
 
 	private ServerSocket serverSocket;
 
-	public static void main(String[] args) throws NumberFormatException, IOException {
-		configureGlobals();
-		ClientProcess process = new ClientProcess(args[0], Integer.valueOf(args[1]));
-	}
-
-	public ClientProcess(String ip, Integer port) {
-		ipAddress = ip;
-		this.port = port;
+	public ClientProcess() {
+		ipAddress = Globals.siteIpAddresses.get(Globals.mySiteId);
+		this.port += Globals.mySiteId;
 		try {
 			BufferedReader cin = new BufferedReader(new InputStreamReader(
 					System.in));
@@ -99,37 +94,4 @@ public class ClientProcess {
 
 		return response;
 	}
-
-	private static void configureGlobals() throws NumberFormatException, IOException {
-		// Read from file and get side ID
-		String path = "config.txt";
-		FileReader fr = new FileReader(path);
-		BufferedReader br = new BufferedReader(fr);
-		
-		int myID = Integer.parseInt(br.readLine());		// line 1
-		String site1 = br.readLine();					// line 2
-		String site2 = br.readLine();					// line 3
-		String site3 = br.readLine();					// line 4
-		String site4 = br.readLine();					// line 5
-		String site5 = br.readLine();					// line 6
-
-
-		site1 = site1.substring(0, site1.indexOf(' '));
-		site2 = site2.substring(0, site2.indexOf(' '));
-		site3 = site3.substring(0, site3.indexOf(' '));
-		site4 = site4.substring(0, site4.indexOf(' '));
-		site5 = site5.substring(0, site5.indexOf(' '));
-		
-		br.close();
-		fr.close();
-		
-		// Set globals
-		Globals.mySiteId = myID;
-		Globals.siteIpAddresses.add(site1);
-		Globals.siteIpAddresses.add(site2);
-		Globals.siteIpAddresses.add(site3);
-		Globals.siteIpAddresses.add(site4);
-		Globals.siteIpAddresses.add(site5);
-	}
-
 }
